Fix incorrect cast in OpenSSL provider
A numeric cast was added to the call to HMAC that doesn't belong there,
this commit reverts that.

Test-Information:

Unit tests pass on macOS 10.14 and Debian 9

+++ Swiften/Crypto/OpenSSLCryptoProvider.cpp
@@ -108,7 +108,7 @@ namespace {
         unsigned int len = SHA_DIGEST_LENGTH;
         std::vector<unsigned char> result(len);
         try {
-            HMAC(EVP_sha1(), vecptr(key), boost::numeric_cast<int>(key.size()), vecptr(data), boost::numeric_cast<int>(data.size()), vecptr(result), &len);
+            HMAC(EVP_sha1(), vecptr(key), boost::numeric_cast<int>(key.size()), vecptr(data), data.size(), vecptr(result), &len);
         }
         catch (const boost::numeric::bad_numeric_cast&) {
             assert(false);
